updated failing tests
