+++ src/main/resources/application.yml
@@ -121,9 +121,9 @@ logging:
     bio.overture.ego: INFO
 
 # Hibernate SQL Debugging
-spring.jpa.properties.hibernate.format_sql: true
-logging.level.org.hibernate.SQL: DEBUG
-logging.level.org.hibernate.type.descriptor.sql: TRACE
+#spring.jpa.properties.hibernate.format_sql: true
+#logging.level.org.hibernate.SQL: DEBUG
+#logging.level.org.hibernate.type.descriptor.sql: TRACE
 
 # When you are desperate, use this...
 #logging.level.org.hibernate: TRACE
@@ -193,7 +193,9 @@ spring:
 ###############################################################################
 spring:
   profiles: test
-logging.test.controller.enable: true
+
+# To log the httpRequests and httpResponses for the controllers, set to true
+logging.test.controller.enable: false
 
 
 spring.datasource:
