+++ src/test/java/bio/overture/ego/controller/ListGroupPermissionsForPolicyControllerTest.java
@@ -1,11 +1,9 @@
 package bio.overture.ego.controller;
 
 import bio.overture.ego.AuthorizationServiceMain;
-import bio.overture.ego.model.dto.PermissionRequest;
 import bio.overture.ego.model.entity.Policy;
-import bio.overture.ego.model.entity.User;
-import bio.overture.ego.model.entity.UserPermission;
-import bio.overture.ego.model.enums.AccessLevel;
+import bio.overture.ego.model.entity.Group;
+import bio.overture.ego.model.entity.GroupPermission;
 import bio.overture.ego.utils.EntityGenerator;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.runner.RunWith;
@@ -16,8 +14,8 @@ import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.TestExecutionListeners;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.context.support.DependencyInjectionTestExecutionListener;
+import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Collection;
 import java.util.List;
 import java.util.UUID;
 
@@ -30,7 +28,8 @@ import static java.lang.String.format;
 @SpringBootTest(
     classes = AuthorizationServiceMain.class,
     webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class UserPermissionControllerTest2 extends AbstractPermissionControllerTest2<User, UserPermission> {
+public class ListGroupPermissionsForPolicyControllerTest extends
+    AbstractListOwnerPermissionsForPolicyControllerTest<Group, GroupPermission> {
 
   @Autowired
   private EntityGenerator entityGenerator;
@@ -39,41 +38,18 @@ public class UserPermissionControllerTest2 extends AbstractPermissionControllerT
   private boolean enableLogging;
 
   @Override
-  protected List<User> setupOwners(String... names) {
-    return entityGenerator.setupUsers(names);
-  }
-
-  @Override
-  protected void createPermissionsForOwners(UUID policyId, AccessLevel mask, Collection<User> owners) {
-    owners.forEach(u ->
-        initStringRequest()
-            .endpoint(getAddPermissionsEndpoint(u.getId()))
-            .body(List.of(PermissionRequest.builder()
-                .mask(mask)
-                .policyId(policyId)
-                .build()))
-            .postAnd()
-            .assertOk());
-  }
-
-  @Override
-  protected Class<User> getOwnerType() {
-    return User.class;
-  }
-
-  @Override
-  protected Class<UserPermission> getPermissionType() {
-    return UserPermission.class;
+  protected List<Group> setupOwners(String ... names) {
+    return entityGenerator.setupGroups(names);
   }
 
   @Override
   protected String getAddPermissionsEndpoint(UUID ownerId) {
-    return format("users/%s/permissions", ownerId);
+    return format("groups/%s/permissions", ownerId);
   }
 
   @Override
   protected String getOwnersForPolicyEndpoint(UUID policyId) {
-    return format("policies/%s/users", policyId);
+    return format("policies/%s/groups", policyId);
   }
 
   @Override
